Merged PR 11764: Restore ability to run Filter service in local containers
This is a pre-work task to the larger performance story coming later. We needed the ability to run Filter service in local containers which was lost as part of the K8 migration.

The solution shown here is taking what we had and making it work; it's not what I'm comfortable with for a 'finished' solution as we still have too many scripts and docker files lying around. You'll see I've condensed several scripts into one and moved away from batch files. The net result is the same outcome as before, Filter svc, it's SQL database and Kafka all running in containers on the localhost.

A rough spec from @<Josh Nimmo> was to reuse Dockerfile files where possible; to avoid a proliferation of them which is fair, and I've tried to satisfy this.

Related work items: #77557