Merged PR 11812: Removed WGSPoint3D class in favor of WGSPoint
For a while we've had two WGS Point classes, the original 'radian' version that existed in 3DP, and the newer 'degrees' version, from what was the Models package. Since then the former has been marked obsolete in favor of the latter.

This piece of work removes the WGSPoint3D class in favor of the newer WGSPoint, degrees orientated version.

It actually wasn't that hard, we did have fears there would be lots of worrisome conversions that needed tidying, but in reality very little conversion of degrees to radians or vice versa takes place. In most cases the WGSPoint is received as a double value on a request, passed to Raptor, and returned in some other form. Whether the owner object thinks of these lat and lon values as degrees or radians is largely moot to the way 3DP acts as a proxy to Raptor.

Related work items: #77936