Merged PR 11801: Data Ocean Client and use of it in Project Service
DataOcean is a replacement for TCC for storing imported files and DXF tiles.
Contract for DataOceanClient is same as for TCCFileAccess. I've only implemented what is required so far, not all functionality from TCFileAccess.
Used in project service where we store and delete files. At the moment imported files are in both TCC and DataOcean. Later we can delete the TCC code.
Added an interface to GracefulWebRequest so we can mock it for unit tests (Steve's idea).

Note the deleted files for the old HttpClient as a copy of the TRex one and the old DataOceanClient under the project service source should never have made it to master. I don't know how that happened. I'm now using GracefulWebRequest as per the rest of our services.

Related work items: #76835