Merged PR 11844: #77828 - Cached items causing 403 due to a non-task cache item being added
BaseProxy Caching layer has two routes to cache items.
The first method adds a Task<T> to the cache given a key. (See BaseProxy.cs:304)
The second method added a T to the cache given a key. (See BasseProxy.cs:318)

If we happen to hit the second method, adding a T to the cache, the first method breaks as it expects a Task<T> to await.

This fix ensures both routes to adding a cache add Task<T>

*Will add some unti tests once nuget packages builds (waiting on aarons PR)*

Related work items: #77828