Merged PR 11848: Add filter validation check to v2 Cell Datum endpoint
As a result of disabling response caching @<Josh Nimmo> found one of the Cell Datum tests was then failing, receiving a response code 200 instead of the expected 400.

This highlighted the issue that the filter UID passed to the endpoint, while a valid Guid, didn't match a filter in the 'mock' database. Because of the nature of the 'GetFilter()' logic it simply returns null in this instance and when the executor receives a null Filter it dutifully passes that to Raptor, resulting in a valid, if not expected, response.

Because v2 endpoints in 3DP in a sense are defined by their filter requirement I've modified the `GetCompactionFilter()` method to take an optional `bool filterMustExist` parameter and for the Cell Datum controller set this to true. This brought the test back into alignment.

It raises a larger issue; should all v2 endpoints that receive a filter UID do the same forced requirement? Generally speaking if the UI gives us a filter uid it is valid so we don't fall into this null filter edge case. But the hole exists all the same. (note, the language here, if the endpoint receives a filter uid but it doesn't resolve to a filter; this is different from receiving a null uid.)

Related work items: #77948