Merged PR 11951: Performance improvement for the creation of transient filters
The goal of this story was to improve batch filter creation performance, which we did. Timings on local development environment dropped from 7+ seconds to create 38 filters in a batch to ~2.7 seconds.

During investigation of this I found we were incorrectly pushing transient filters to Kafka (wrong) and not persistent (wrong), so flipped this behaviour at the same time.

Related work items: #77557