Merged PR 11952: #78193 - Files uploaded via Project Service are being corrupted when they are
When uploading files via FileAccess to TCC, the GracefulWebRequest was taking the binary stream and assuming it was as string.

To resolve, TCC now passes a content type and GracefulWebRequest does not do any further encoding of the streams (as the stream is already an encoded string if the data is in fact a string).

Related work items: #78193