Merged PR 11982: Fix elevation palette generation so it returns 31 color items
The fix made in #75449 removed the below and above colors from the elevation palette. Unfortunately this reduced the number of array elements from 31 to 29. Apparently all palettes in 3DP contain 31 elements and the UI relies on this assumption.

This story is to put back the two missing elements but retain the color palette, e.g. without the pink and purple below and above colors seen in the earlier #75449.

It turns out the action method was assuming the responsibility of truncating the color palette and removing the first and last elements. So I targeted this in the fix.

Related work items: #78267