Merged PR 12049: Refactor ConvertFilter() method to tidy code and remove technical debt
This was some technical debt I undertook over the holiday period.

The `ConvertFilter` method was split (from a legacy decision to do filter resolution in the executor) and also contained no longer used parameters. I tidied these up merging the two methods and removing the unused parameters. This affected every call to `ConvertFilter`. Also added unit tests for the method. It's not 100% coverage, but a good swathe in lieu of what we had before.

Related work items: #78024