Merged PR 12075: Fix transiently failing Pass Count Summary test
For a long time we've been plagued by a transiently failing acceptance test in 3DP. The cause has finally been identified and today, fixed!

It would occur sometimes because of the way Raptor nodes process pass count data (no error here). They would record the min and max pass count target values seen when processing cell data. If multiple values were found, e.g. the minimum passcount target for one tag file was 3, and another was 5, it would set the `targetVaries` flag to true. The min and max values in the `PassCountTarget` object would reflect 'one' of the passes from one of the Raptor nodes and so be irrelevant when targetVaries=true.  Our acceptance tests were treating these min and max values as gospel as part of the response validation.

Example response body:

      "PassCountTarget": {
        "minPassCountMachineTarget": 5.0,
        "maxPassCountMachineTarget": 5.0,
        "targetVaries": true
      }

The fix was actually for Raptor to zero the min and max fields if targetVaries=true. The ultimate fix in this PR is the change to the Pass Count response JSON file for six response objects.

This doesn't affect the UI because it shows either 'varies' if targetVaries=true, or the actual min/max values if false. So the zero'ing of the min and max when true should not affect the UI.

Unfortunately there are other code changes, tidying and refactoring in this PR; they occurred through both experimentation in tracking down the issue and a little bit of general tidying as I went. They can be largely summarized as:

1. async await changes to improve concurrency.
2. In the summary controller made changes so `GetCompactionFilter()` is only called once per request; was usually being called twice due to it's placement in two validation methods.
3. Improved mock web api performance by setting up singleton service classes to provide the static test data; in this way it's not re created for each request.
4. Introduced the concept of using local functions in async methods where there are eager preconditions; some reading if you're interested, https://blogs.msdn.microsoft.com/seteplia/2017/10/03/dissecting-the-local-functions-in-c-7/ and https://blogs.msdn.microsoft.com/seteplia/2018/01/25/the-performance-characteristics-of-async-methods/.
5. Lots of CTRL+K+D, and CTRL+R+G (format document and sort and remove usings respectively). Sorry Elspeth.

Related work items: #78217