Merged PR 12130: Introduce API to extract boundaries as GeoJson from DXF linework file
This PR is missing some functional requirements but needs to be deployed into DEV so the UI team can complete their development phase. It needs to be released today though.

What it does now:
Given a specific filename (file is known to exist in a folder on the IO node) on the request (GET /api/v2/linework/boundaries) the service will process and return the boundaries in GeoJson format. This phase includes uploading the provided file.

What it needs to do:
  1. (Last functional step) It needs to upload the file to the IO node for processing by Raptor (boundary extraction). It's currently uploading to AS node.
  2. It needs more logging.
  3. It needs the acceptance tests to be completed.
  4. It (may) need configuration settings added to set the upload and working directories.

What the UI team will be given with the feature as is:
  1. a test DXF file to upload
  2. a valid filename that is present on the IO node working directory
  3. the contract definition for /api/v2/linework/boundaries that will include using the file and filename provided above

Related work items: #77252