Merged PR 12186: Stage 3 / 3 Cache fixes - add notifications from Project service, and Cache Invalidation cache that listens for the notifications
This change adds a CacheInvalidationService in the Push Service Notifications (I couldn't think of a better place, but open to suggestions).

If the CacheInvalidationSerivce is registered via DI it will listen for Project / Customer / User change notifications, and invalidate any cache related to the model changed.
Current only 3dp / project service has this enabled - we can enable as needed to other services, but that will be another story.

I've modified the Project service to push notifications when things change - no other services as of yet.

Related work items: #78314