Merged PR 12354: Make DxfUnits parameter mandatory for linework requests
As a developer I want to make the `DxfFileRequest::DxfUnits` field mandatory so it's not possible to send no value and default to a unit of measurement that may or may not work with the linework file.

The change meant that the acceptance test couldn't POST a serialized concrete type because it would naturally inherit the behaviour of treating DxfUnits as null and therefore pass a zero int to the API.
Changed acceptance tests to use JObject for POST request objects so we can simply serialize into them the JSON read from static request files.

Related work items: #79375