Merged PR 12375: #79405 - First pass at master data endpoint (it's slow, but it works for now)
This is slow, we iterate many times over the master data models and execute a new request against the end points.
The second hit is a lot faster, due to cache - but this will be improved in due time (once we're happy with the data we are getting back).

Related work items: #79405