Merged PR 12388: Invalidate project cache when creating a persistent file that contains an alignment file or design boundary
Problem description: If you add a persistent filter that uses a design boundary or an alignment file. Save it then go out to the dashboard. Go into project files and delete the file that is in the filter. This allows the file to be deleted.

The fix is to force the Project's cache item to be invalidated under these filter creation conditions. Then when the project service next fetches the file it'll be informed it's used as part of a filter and deletion is not allowed.

The trick was bubbling the flag back up to the controller from the executor; which was the easiest and most logical place to determine if we're dealing with a design boundary or alignment file.

Note, this piece of work includes changing `startup.cs` to inherit from `BaseStartup` as per the new pattern.

Please confirm; is my check on `ContainsBoundary` correct? It's OK to include `PolygonName`?

Related work items: #79370