Merged PR 12428: DataOcean and Pegasus
Two libraries one for DataOcean and one for Pegasus. DataOcean is the replacement for TCC i.e. where files are stored. Pegasus is the engine for generating tiles to replace GlobalMapper. The DataOcean library interface is approx the same as TCCFileAccess.

For create project, the coordinate system file is saved to DataOcean.
For create/update imported file, the UI calls the background upload in project service, which schedules the job with the scheduler. The scheduler calls back to the project service on the internal end point which goes through the whole process of database update, saving to TCC, notifying Raptor etc. I have added saving the file to DataOcean. For an alignment file, it also calls 3dpm to get the generated DXF file through Raptor which it saves to DataOcean. For now, due to timing issues with web api calls open for a long time, the actual DXF tile generation has been removed from the project service. (It was part of the call back process above where project service called tile service proxy to do the generation using Pegasus).

We (Dmitry.Steve, Elspeth) are reviewing how/where to do the dxf tile generation, probably the scheduler. The work will be done on a separate branch.

Related work items: #76835