Merged PR 12571: Fixed setting of KAFKA_ADVERTISED_HOST_NAME environment variable (in local testing)
As a developer I need the local 'build and run docker' scripts to correctly set the KAFKA_ADVERTISED_HOST_NAME environment variable so Kafka doesn't error trying to find the node.

@<Elspeth Jeet> found this originally a couple of weeks ago and we fixed it locally for her branch. But the wider issue still remained.

The container running Kafka needs the IP address of the host (in our case) for it's host name, if that's not set, or not found, it fails to work correctly, affecting the outcomes of running the owner service locally in containers.

This quick fix resolves the issue.

Also tweaked scripts to not be so overtly aggressive in removing stopped containers (now only targets related service containers) and definitely doesn't delete images any longer.

Related work items: #80235