Merged PR 12518: Change boundaries API to use multipart/form-data requests
The goal of this PR is to address a couple of workflow errors in the original logic that uses Raptor to extract boundaries from DXF linework files.

We need to pass the coordinate system file (.dc) so it's available to Raptor when parsing the DXF file. It looks for it in the same folder.
Also need to not pass the Canary project id, but -1, to indicate no project. This trips the following rules:

`If project id = -1 and file descriptor includes a path then use that path to locate the named DXF and DC files and extract boundaries using these inputs.`

Ordinarily it's sufficient to pass just the project id and a DXF filename; the DC file information will be pulled from the existing project.
We don't have a project to work from in this scenario.

Also expanded the acceptance tests to support multipart/form-data requests; this isn't generic yet, but will become so as need dictates. The capability should be rolled into the existing RestClient class.

Related work items: #79907