Merged PR 12632: DXF tile generation through Scheduler job
DXF tile generation is done through the VSS jobs general mechanism set up in the scheduler (as discussed at the whiteboard meeting).
Project service calls scheduler to put the job in the queue.
Scheduler runs the job which uses Pegasus. If it fails a Slack notification is sent.
If it succeeds a push service notification is sent which is picked up by the project service. The project service updates the database with the zoom levels.

Related work items: #76835, #79758