Merged PR 12757: #80223 - Add service resolvers for kubernetes and configuration, and a service
This is the first pass for a ServiceDiscovery class for our services.

The new class uses both Kubernetes and IConfiguration to attempt to find services.
For Kubernetes, we need to add a label to our Services to indicate the service name.

The proxy classes will need to change quite a bit to use the new service discovery, as the service discovery will not add any endpoints to the url
e.g http://localhost/models/ is not a valid service endpoint, only http://localhost would be returned. Therefor the proxy would need to know about the route.

For example: _RAPTOR_NOTIFICATION_API_URL: https://api-stg.trimble.com/t/trimble.com/vss-dev-3dproductivity/2.0/notification_ is invalid using the new service discovery.

Note: We will need to add a few configuration values to all services as well, but we will do that as we migrate to the new proxies (which will be next)

Related work items: #80223