Merged PR 12790: Review and tweak checking of ProjectId on executor calls to Raptor
The idea behind this was to review all the places in 3DP executors where we do `request.ProjectId ?? -1` and remove the null check because only one endpoint in Raptor allows a project Id of `-1`.

However, it's a messy situation because `ProjectId` is a nullable long and for reasons I won't go into here; it's not a simple matter of removing the nullable component and then removing these null coalescing checks.

So all I've done here is;

1. Moved the `VelociraptorConstants` class to a more accessible 3DP project so it's more widely available.
2. Replaced occurrences of `-1` for project Id setting to use a constant.
3. Removed the redundant projectId and projectUid check from the `ProjectID` class. They're already being validated by the attribute validation method in `ProjectID::Validate()`.
4. Consolidated nuget packages for 3DP.

Related work items: #78023