Merged PR 12830: Add 1 hour tollerance for Pass Count and Veta export report tests
As a developer I don't want to re create our export report test data each time North America goes through a DST change so I need the test runner to be smarter about subtle differences.

This piece of work fixes the affected tests so they now have a 1 hour tolerance on the `Time` column in the CSV export report. Clearly this introduces the possibility we won't detect actual faults in time data that are within an hour of what we're expecting. The function could be improved to act only if DST is in play in the affected project's timezone. This is beyond the scope of this 'quick fix' to unblock the 3DP build. It may be agreed what's done here is sufficient.

Related work items: #81106