Merged PR 12832: Upgrad package VSS.Velociraptor.Interfaces from 1.0.76 to 1.0.78
Upgraded package VSS.Velociraptor.Interfaces from 1.0.76 to 1.0.78 and fixed breaking change.

I needed this piece of work done to progress with the "CSIB from project id" piece of work for the data migration story. I chose to do it separately because the shim package included a breaking change in 3DP (bool became enum) and to not block others working in 3DP pushed it through as a separate piece of work.

Also included:
* Very minor code tidying in acceptance tests; to improve clarity when reading failed test logging in VSTS
* Pushed the tolerance on another transiently failing test.

Related work items: #81103