Merged PR 12841: Add API to return the CSIB for a project
As a developer trying to migrate from TCC to Data Ocean I need a way to get the DC file for all projects being processed.

This story adds a new controller/endpoint to the 3DP service that simply returns (from Raptor) the CSIB for a project. This is necessary for the Data Ocean migration as many projects don't have a valid DC file any longer and we don't have the appropriate DC file on TCC. Using this endpoint and then passing that CSIB to the coordinate service gives us a guaranteed DC file for the project. Necessary for Pegasus to gen tiles in Data Ocean.

Related work items: #81020