Merged PR 12808: #80223 Update helm charts to include service name ***NO_CI***
Updated helm charts for all of services required for Service Discovery

Also, had to change the kubernetes logic when attempting to discover services to look for a TCP port listening on 80, rather than an HTTP port listening on anything. As our services are defined as TCP port 80.
We could change this back, but we would need to make changes to the kubernetes service definitions

Related work items: #80223