Merged PR 12991: Add kubernetes namespace to configmaps
Added code to allow kubernetes to detect the name space, but it didn't work in all cases.
Added kubernetes namespace to config maps, which works for now - but if we change namespaces per application later on, we'll need to reevaluate

Added IsEnabled flag to Service Resolvers - so they aren't checked if not needed (if the resources aren't available).

Confirmed working in Dev using 3dnow and Kubernetes service resolver

Related work items: #80223