Merged PR 13075: Add global .editorconfig for the mono repository
A global .editorconfig allows us to create a consistent uniform coding style ruleset that's visible and applicable to all projects in the mono repo despite the developers use of R# or not.

See attached clip for an example of how the settings are changed and reflected in Visual Studio.

TRex, 3DP and Filter service already used one, I've pushed that file to the root of the repo and added it as a solution item to other solution files.

**Important:** This story only deals with adding a very generic .editorconfig that doesn't affect the rules we've already agreed to. You're not going to be bombarded with editor suggestions or warnings that weren't there yesterday.

To watch this .gif multiple times right click on the following image and select 'Open image in new tab'.

![editor-config-demo.gif](https://visionlink.visualstudio.com/badf5c55-c8a5-46da-8e96-74337945e1f3/_apis/git/repositories/ba0f8953-a501-41fd-8a26-0947e45dceab/pullRequests/13075/attachments/editor-config-demo.gif)

Related work items: #81632