Merged PR 13564: #83159 - Move IConfigurationStore to Common Abstractions in prep for other
This PR should only be to move `IConfigurationStore` from it's own Project to `VSS.Common.Abstractions.Configuration`.

This is prep work for the actual task, where an interface needs to be injected with `IConfigurationStore`, but the interface needs to be in common (which the `ConfigurationStore` implementation references - something something circular reference).

There should be **NO** logic changes in this PR

Related work items: #83159