Merged PR 13599: #81222 Change Dot on Map model and add tests
Change the dot on the map model that is sent via SignalR to a more streamlined model.

This moves the business logic to the backend rather than UI.

The model is extendable, but only implements a few fields for now until we finalize the design.

Added some unit tests to ensure the logic is respected too

Related work items: #81222