Merged PR 13612: #81222 - Fix up incorrect caching of Assets
Subtle issue in the caching on Assets, where we didn't generate a unique cache key - meaning every asset matched to the same result.
The cache key is built up from the Uid and the UserId provided to the base proxy, which were null for the asset proxy calls - meaning the cache key became the name of the Model being returned, regardless of any parameters passed in.

Fixed the Proxy, and the cache will now log this and prevent any other cases

Related work items: #81222