Merged PR 13804: DDV Model Updates
The DDV Models for V1 endpoint are not using the latest naming convention, and contain magic numbers.

This change fixes that, but converting magic numbers to null, and tidying up the model names.

Sadly, the only real good way to solve this is to copy the models and slightly modify them :( `CellPassesV2Result` is a direct copy of the v1 result, with names updated.

Related work items: #83315