Merged PR 13888: #81670 - Aws script was not copying correctly.
#81670 - Aws script was not copying correctly - therefor route wasn't getting created

Forgot to copy the AWS Route script to the docker image, therefor the route wasn't adding.
I've rebuilt a new image, and validated that the route created is correct.

Confirmed the script is run at application startup as seen here:
![image.png](https://visionlink.visualstudio.com/badf5c55-c8a5-46da-8e96-74337945e1f3/_apis/git/repositories/ba0f8953-a501-41fd-8a26-0947e45dceab/pullRequests/13888/attachments/image%20(2).png)

And we can query the AWS Meta Data service correctly (previously this would fail due to no route to the server)
![image.png](https://visionlink.visualstudio.com/badf5c55-c8a5-46da-8e96-74337945e1f3/_apis/git/repositories/ba0f8953-a501-41fd-8a26-0947e45dceab/pullRequests/13888/attachments/image%20(3).png)

Related work items: #81670