Merged PR 14123: Geotiff backed integration
This PR request completes integration of Geotiff into the back end services.
I have also tidied up some old code but decided not to proceed with refactoring as 2/3 of the existing code will eventually die as we move over to TRex and Pegasus. I think we will need to tidy up that code as we cross those bridges.

Related work items: #84882