Merged PR 14327: Add webapi component to helm charts
We found a situation with the project service where the consumer would be given queries for webapi and (rightly) respond with connection refused. This is because the service was using the app label selector in the helm chart to determine the endpoints for the webapi.

This introduces a component label of webapi and uses this to select the webapi components.  This pattern is applied to services which have pods/containers which are not webapi.

This will also fix some issues relating to long branch names not deploying.

Related work items: #85960