Merged PR 14428: Common Date fixes
#Summary
- This fixes the case where we have a local date but want the Iso8601DateTimeString, this change allows that to be obtained.
- Fixes to helm chart to better handle long or uppercase chars in branch names

##More info

When string dates such as "2019-07-30T23:35:12Z" are constructed using [FromQuery] in controllers they become "Local Dates" they have the correct time,  and we can get utc value of these (i.e. the original time passed in) by using the ToUniversalTime().

In the case when we have a string with no timezone e.g. "2019-07-30T23:35:12" (No Z suffix) then this becomes an unknown type and we assume utc was intended.

Related work items: #86173