Merged PR 15242: Scheduler Refactoring as well as automated pass export
This is driven by #88852  - please read through it.

For now, customers and emails are hardcoded as i want to get the initial feedback and not spoil the config.

It is also addressing some technical debt in the scheduler specifically a "special" export job is now using the standard pattern.

Some minor improvements in logging and exceptions as well.

Related work items: #88743, #88852