Merged PR 551: Speed up project list, with has more, 200 projects were timing out on tpaas -
We need to speed up project list - eventually we will change the API to return only what is needed - but that is extra work right now (e.g list projects for UI only needs some - and not all the details, but list for validations needs more including all details).

Eventually list projects for UI could just return the summary data, and get single project with all data for the map view - but we need to change the project service api for that.

This just runs the get project details in parallel (sometimes we need a separate call to populated the boundary values) - which makes the API run a lot faster